we now can open the "content.html" entry in a browser
This still isn't quite optimal. We might want to be able to just view
the raw html of the page instead of this whole redirect business. It
shouldn't be too hard to just get the actual page contents and name it
content.html and have something called redirect.html that does the
redirect.